src/IndexingFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

openai
azure-functions
azure.storage.blob
azure-search-documents
azure.identity
llama-index
llama-index-vector-stores-azureaisearch
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-readers-azstorage_blob
pyyaml